+++ src/api/account.ts
@@ -214,7 +214,7 @@ export class Account {
     await waitForIndexerOnVersion({
       config: this.config,
       minimumLedgerVersion: args.minimumLedgerVersion,
-      processorTypes: [ProcessorType.ACCOUNT_TRANSACTION_PROCESSOR],
+      processorType: ProcessorType.ACCOUNT_TRANSACTION_PROCESSOR,
     });
     return getAccountTokensCount({
       aptosConfig: this.config,
@@ -244,7 +244,7 @@ export class Account {
     await waitForIndexerOnVersion({
       config: this.config,
       minimumLedgerVersion: args.minimumLedgerVersion,
-      processorTypes: [ProcessorType.TOKEN_V2_PROCESSOR],
+      processorType: ProcessorType.TOKEN_V2_PROCESSOR,
     });
     return getAccountOwnedTokens({
       aptosConfig: this.config,
@@ -276,7 +276,7 @@ export class Account {
     await waitForIndexerOnVersion({
       config: this.config,
       minimumLedgerVersion: args.minimumLedgerVersion,
-      processorTypes: [ProcessorType.TOKEN_V2_PROCESSOR],
+      processorType: ProcessorType.TOKEN_V2_PROCESSOR,
     });
     return getAccountOwnedTokensFromCollectionAddress({
       aptosConfig: this.config,
@@ -306,7 +306,7 @@ export class Account {
     await waitForIndexerOnVersion({
       config: this.config,
       minimumLedgerVersion: args.minimumLedgerVersion,
-      processorTypes: [ProcessorType.TOKEN_V2_PROCESSOR],
+      processorType: ProcessorType.TOKEN_V2_PROCESSOR,
     });
     return getAccountCollectionsWithOwnedTokens({
       aptosConfig: this.config,
@@ -328,7 +328,7 @@ export class Account {
     await waitForIndexerOnVersion({
       config: this.config,
       minimumLedgerVersion: args.minimumLedgerVersion,
-      processorTypes: [ProcessorType.ACCOUNT_TRANSACTION_PROCESSOR],
+      processorType: ProcessorType.ACCOUNT_TRANSACTION_PROCESSOR,
     });
     return getAccountTransactionsCount({
       aptosConfig: this.config,
@@ -357,7 +357,7 @@ export class Account {
     await waitForIndexerOnVersion({
       config: this.config,
       minimumLedgerVersion: args.minimumLedgerVersion,
-      processorTypes: [ProcessorType.FUNGIBLE_ASSET_PROCESSOR],
+      processorType: ProcessorType.FUNGIBLE_ASSET_PROCESSOR,
     });
     return getAccountCoinsData({
       aptosConfig: this.config,
@@ -379,7 +379,7 @@ export class Account {
     await waitForIndexerOnVersion({
       config: this.config,
       minimumLedgerVersion: args.minimumLedgerVersion,
-      processorTypes: [ProcessorType.FUNGIBLE_ASSET_PROCESSOR],
+      processorType: ProcessorType.FUNGIBLE_ASSET_PROCESSOR,
     });
     return getAccountCoinsCount({ aptosConfig: this.config, ...args });
   }
@@ -414,7 +414,7 @@ export class Account {
     await waitForIndexerOnVersion({
       config: this.config,
       minimumLedgerVersion: args.minimumLedgerVersion,
-      processorTypes: [ProcessorType.FUNGIBLE_ASSET_PROCESSOR],
+      processorType: ProcessorType.FUNGIBLE_ASSET_PROCESSOR,
     });
     return getAccountCoinAmount({ aptosConfig: this.config, ...args });
   }
@@ -437,7 +437,7 @@ export class Account {
     await waitForIndexerOnVersion({
       config: this.config,
       minimumLedgerVersion: args.minimumLedgerVersion,
-      processorTypes: [ProcessorType.DEFAULT],
+      processorType: ProcessorType.DEFAULT,
     });
     return getAccountOwnedObjects({
       aptosConfig: this.config,